coding days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2757 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/Grid.hpp
@@ -51,7 +51,7 @@ struct RegularGridConfiguration
 	/// number of dimensions
 	size_t dim_;
 	/// number of levels
-	size_t level_;
+	int level_;
 };
 
 /**
